refactor: separate logger package
