+++ agent_test.go
@@ -1,9 +1,6 @@
 package dagu
 
 import (
-	"bytes"
-	"io"
-	"log"
 	"net/http"
 	"net/url"
 	"os"
@@ -307,44 +304,4 @@ func testDAGAsync(t *testing.T, file string) (*Agent, *controller.DAG) {
 	}()
 
 	return a, dag
-}
-
-func TestTeeLogger(t *testing.T) {
-	origStdout := os.Stdout
-	r, w, err := os.Pipe()
-	require.NoError(t, err)
-	os.Stdout = w
-	log.SetOutput(w)
-
-	defer func() {
-		os.Stdout = origStdout
-		log.SetOutput(origStdout)
-	}()
-
-	tmpDir := utils.MustTempDir("test-tee-logger")
-	tmpFile := path.Join(tmpDir, "test.log")
-	f, err := os.Create(tmpFile)
-	require.NoError(t, err)
-	l := &teeLogger{File: f}
-	err = l.Open()
-	require.NoError(t, err)
-
-	text := "test log"
-	log.Println(text)
-	os.Stdout = origStdout
-	_ = w.Close()
-	l.Close()
-
-	var buf bytes.Buffer
-	_, err = io.Copy(&buf, r)
-	require.NoError(t, err)
-
-	s := buf.String()
-	require.Contains(t, s, text)
-
-	f, err = os.Open(tmpFile)
-	require.NoError(t, err)
-	b, err := io.ReadAll(f)
-	require.NoError(t, err)
-	require.Contains(t, string(b), text)
 }
